# boleto-api-python/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Boleto-API (Python)",
    "description": "Gateway de cobrança multi-banco (C6/Sicoob) + proxy ao engine brcobrança (Ruby).\n\n**Produto standalone** consumido por múltiplos sistemas; escopo por `tenant_id`.\n\n**Push de eventos (saída, não é um path desta API):** ao receber o webhook do banco, o gateway envia o evento normalizado (`WebhookEvent`) por `POST` ao consumidor dono do tenant, assinado em `X-Signature: sha256=<hmac_sha256(secret, raw_body)>`. Destino por tenant via `SUB__<tenant>__URL/SECRET`, com fallback global `EVENT_WEBHOOK_URL/SECRET`.\n\n**Respostas `dict` (passthrough):** rotas Pix/Pix Automático/extrato/conciliação devolvem o corpo do banco como veio (padrão BACEN/documentação do banco), sem re-tipagem — o contrato é o do banco; o gateway normaliza autenticação, multi-tenant e erros.",
    "version": "0.6.0"
  },
  "paths": {
...
    "/conciliacao/recebiveis": {
      "get": {
        "tags": [
          "conciliacao"
        ],
        "summary": "Recebiveis",
        "description": "Recebíveis C6 Pay do período (paginado; máx. 60 dias por consulta).",
        "operationId": "recebiveis_conciliacao_recebiveis_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Data inicial (YYYY-MM-DD)",
              "title": "Start Date"
            },
            "description": "Data inicial (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Data final (YYYY-MM-DD, máx. 60 dias após start)",
              "title": "End Date"
            },
            "description": "Data final (YYYY-MM-DD, máx. 60 dias após start)"
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Provider",
              "default": "c6"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 50,
              "title": "Size"
            }
          },
          {
            "name": "X-Bank-Credentials",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Credenciais do banco (JSON em base64) — só memória, nunca persistidas. Fallback: cofre VAULT__*.",
              "title": "X-Bank-Credentials"
            },
            "description": "Credenciais do banco (JSON em base64) — só memória, nunca persistidas. Fallback: cofre VAULT__*."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer bapi_... (token do /credenciais)",
              "title": "Authorization"
            },
            "description": "Bearer bapi_... (token do /credenciais)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConciliacaoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conciliacao/transacoes": {
      "get": {
        "tags": [
          "conciliacao"
        ],
        "summary": "Transacoes",
        "description": "Transações C6 Pay do período (paginado; máx. 60 dias por consulta).",
        "operationId": "transacoes_conciliacao_transacoes_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Data inicial (YYYY-MM-DD)",
              "title": "Start Date"
            },
            "description": "Data inicial (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Data final (YYYY-MM-DD, máx. 60 dias após start)",
              "title": "End Date"
            },
            "description": "Data final (YYYY-MM-DD, máx. 60 dias após start)"
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Provider",
              "default": "c6"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "default": 50,
              "title": "Size"
            }
          },
          {
            "name": "X-Bank-Credentials",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Credenciais do banco (JSON em base64) — só memória, nunca persistidas. Fallback: cofre VAULT__*.",
              "title": "X-Bank-Credentials"
            },
            "description": "Credenciais do banco (JSON em base64) — só memória, nunca persistidas. Fallback: cofre VAULT__*."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer bapi_... (token do /credenciais)",
              "title": "Authorization"
            },
            "description": "Bearer bapi_... (token do /credenciais)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConciliacaoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConciliacaoOut": {
        "properties": {
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "last_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Page"
          },
          "total_items": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Items"
          },
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "title": "ConciliacaoOut",
        "description": "Página de recebíveis/transações — itens crus do banco + paginação."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Provider": {
        "type": "string",
        "enum": [
          "brcobranca",
          "c6",
          "sicoob"
        ],
        "title": "Provider",
        "description": "Provider de cobrança. `brcobranca` = offline/CNAB (engine Ruby)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}